+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to the extension are documented in this file.
 <!-- Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file. -->
 
 ## [Unreleased]
+- The extension is now logging to a vscode output channel.
 
 ## [0.1.1] - 2022-07-29